Explicitly instantiate array_like
